More cleanup, deleting unused files in this branch
