+++ domain/src/main/java/starter/project/domain/Car.java
@@ -1,11 +1,10 @@
-package starterProject.domain;
+package starter.project.domain;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.ArrayList;
 @JsonDeserialize
 @Data @NoArgsConstructor @AllArgsConstructor
 @Builder public class Car {
