+++ domain/src/main/java/starter/project/domain/Transmission.java
@@ -1,14 +1,16 @@
-package starterProject.domain;
+package starter.project.domain;
+
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 @Data
 @Builder
 @JsonDeserialize
 @NoArgsConstructor
 @AllArgsConstructor
-public class Wheels {
-    private int wheels;
+public class Transmission {
+    private Boolean automatic;
 }
