+++ domain/src/main/java/starter/project/domain/Wheels.java
@@ -1,16 +1,14 @@
-package starterProject.domain;
-
+package starter.project.domain;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 @Data
 @Builder
 @JsonDeserialize
 @NoArgsConstructor
 @AllArgsConstructor
-public class Doors {
-    private int doorNumbers;
+public class Wheels {
+    private int wheels;
 }
