--- domain/src/main/java/starterProject/App.java
@@ -1,31 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package starterProject;
-
-import starterProject.domain.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class App {
-    public String getGreeting() {
-        return "Hello world.";
-    }
-    public static void main(String[] args) {
-        createCar();
-    }
-    private static void createCar() {
-        Transmission automatic = Transmission.builder().automatic(true).build();
-        Wheels parelli = Wheels.builder().wheels(4).build();
-        Engine v12 = Engine.builder().cylinders(12).build();
-        Doors verticalDoors = Doors.builder().doorNumbers(2).build();
-        Car ferarri = Car.builder().transmission(automatic).wheels(parelli).engine(v12).doors(verticalDoors).name("ferrari").build();
-        Car buggatti = Car.builder().transmission(automatic).wheels(parelli).engine(v12).doors(verticalDoors).name("bugatti").build();
-        Car landRover = Car.builder().transmission(automatic).wheels(parelli).engine(v12).doors(verticalDoors).name("landrover").build();
-        List <Car> objectList =  new ArrayList<>();
-        objectList.add(ferarri);
-        objectList.add(buggatti);
-        objectList.add(landRover);
-    }
-}
