+++ domain/src/test/java/starter/project/domain/TransmissionTest.java
@@ -1,0 +1,5 @@
+package starter.project.domain;
+
+public class TransmissionTest {
+
+}
