--- frontend/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.h2.console.enabled=true
-spring.datasource.platform=h2
-spring.datasource.url=jdbc:h2:mem:testdb
